Refactor Emit/Run/Verify test utilities
