+++ src/Test/Utilities/Desktop/CommonTestBase.CompilationVerifier.cs
@@ -45,11 +45,6 @@ namespace Microsoft.CodeAnalysis.Test.Utilities
                 _dependencies = dependencies;
             }
 
-            public CompilationVerifier Clone()
-            {
-                return new CompilationVerifier(_test, _compilation, _dependencies);
-            }
-
             internal CompilationTestData TestData
             {
                 get { return _testData; }
@@ -94,7 +89,7 @@ namespace Microsoft.CodeAnalysis.Test.Utilities
 
             public void Emit(string expectedOutput, IEnumerable<ResourceDescription> manifestResources, bool peVerify, SignatureDescription[] expectedSignatures)
             {
-                using (var testEnvironment = new HostedRuntimeEnvironment(_dependencies))
+                using (var testEnvironment = RuntimeUtilityFactory.NewInstance(_dependencies))
                 {
                     string mainModuleName = Emit(testEnvironment, manifestResources);
                     _allModuleData = testEnvironment.GetAllModuleData();
@@ -120,7 +115,7 @@ namespace Microsoft.CodeAnalysis.Test.Utilities
             // Replace bool verify parameter with string[] expectedPeVerifyOutput. If null, no verification. If empty verify have to succeed. Otherwise compare errors.
             public void EmitAndVerify(params string[] expectedPeVerifyOutput)
             {
-                using (var testEnvironment = new HostedRuntimeEnvironment(_dependencies))
+                using (var testEnvironment = RuntimeUtilityFactory.NewInstance(_dependencies))
                 {
                     string mainModuleName = Emit(testEnvironment, null);
                     string[] actualOutput = testEnvironment.PeVerifyModules(new[] { mainModuleName }, throwOnError: false);
@@ -128,14 +123,14 @@ namespace Microsoft.CodeAnalysis.Test.Utilities
                 }
             }
 
-            private string Emit(HostedRuntimeEnvironment testEnvironment, IEnumerable<ResourceDescription> manifestResources)
+            private string Emit(IRuntimeUtility testEnvironment, IEnumerable<ResourceDescription> manifestResources)
             {
                 testEnvironment.Emit(_compilation, manifestResources);
 
                 _diagnostics = testEnvironment.GetDiagnostics();
                 EmittedAssemblyData = testEnvironment.GetMainImage();
                 EmittedAssemblyPdb = testEnvironment.GetMainPdb();
-                _testData = testEnvironment.GetCompilationTestData();
+                _testData = ((IInternalRuntimeUtility) testEnvironment).GetCompilationTestData();
 
                 return _compilation.Assembly.Identity.GetDisplayName();
             }
