+++ src/Test/Utilities/Desktop/HostedRuntimeEnvironment.cs
@@ -1,24 +1,21 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
 using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Collections.Immutable;
-using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Reflection;
-using System.Runtime.Serialization;
-using System.Text;
 using System.Threading;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CodeGen;
 using Microsoft.CodeAnalysis.Emit;
+using Roslyn.Test.Utilities;
 using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.Test.Utilities
 {
-    public sealed class HostedRuntimeEnvironment : IDisposable
+    public sealed class HostedRuntimeEnvironment : IDisposable, IRuntimeUtility, IInternalRuntimeUtility
     {
         private sealed class RuntimeData : IDisposable
         {
@@ -52,18 +49,6 @@ namespace Microsoft.CodeAnalysis.Test.Utilities
                 }
 
                 Disposed = true;
-            }
-        }
-
-        private struct EmitOutput
-        {
-            internal ImmutableArray<byte> Assembly { get; }
-            internal ImmutableArray<byte> Pdb { get; }
-
-            internal EmitOutput(ImmutableArray<byte> assembly, ImmutableArray<byte> pdb)
-            {
-                Assembly = assembly;
-                Pdb = pdb;
             }
         }
 
@@ -200,200 +185,6 @@ namespace Microsoft.CodeAnalysis.Test.Utilities
             }
         }
 
-        /// <summary>
-        /// Find all of the <see cref="Compilation"/> values reachable from this instance.
-        /// </summary>
-        /// <param name="compilation"></param>
-        /// <returns></returns>
-        private static List<Compilation> FindReferencedCompilations(Compilation original)
-        {
-            var list = new List<Compilation>();
-            var toVisit = new Queue<Compilation>(FindDirectReferencedCompilations(original));
-
-            while (toVisit.Count > 0)
-            {
-                var current = toVisit.Dequeue();
-                if (list.Contains(current))
-                {
-                    continue;
-                }
-
-                list.Add(current);
-
-                foreach (var other in FindDirectReferencedCompilations(current))
-                {
-                    toVisit.Enqueue(other);
-                }
-            }
-
-            return list;
-        }
-
-        private static List<Compilation> FindDirectReferencedCompilations(Compilation compilation)
-        {
-            var list = new List<Compilation>();
-            var previousCompilation = compilation.ScriptCompilationInfo?.PreviousScriptCompilation;
-            if (previousCompilation != null)
-            {
-                list.Add(previousCompilation);
-            }
-
-            foreach (var reference in compilation.References.OfType<CompilationReference>())
-            {
-                list.Add(reference.Compilation);
-            }
-
-            return list;
-        }
-
-        /// <summary>
-        /// Emit all of the references which are not directly or indirectly a <see cref="Compilation"/> value.
-        /// </summary>
-        private static void EmitReferences(Compilation compilation, HashSet<string> fullNameSet, List<ModuleData> dependencies, DiagnosticBag diagnostics)
-        {
-            // NOTE: specifically don't need to consider previous submissions since they will always be compilations.
-            foreach (var metadataReference in compilation.References)
-            {
-                if (metadataReference is CompilationReference)
-                {
-                    continue;
-                }
-
-                var peRef = (PortableExecutableReference)metadataReference;
-                var metadata = peRef.GetMetadata();
-                var isManifestModule = peRef.Properties.Kind == MetadataImageKind.Assembly;
-                var identity = isManifestModule
-                    ? ((AssemblyMetadata)metadata).GetAssembly().Identity
-                    : null;
-
-                // If this is an indirect reference to a Compilation then it is already been emitted 
-                // so no more work to be done.
-                if (isManifestModule && fullNameSet.Contains(identity.GetDisplayName()))
-                {
-                    continue;
-                }
-
-                foreach (var module in EnumerateModules(metadata))
-                {
-                    ImmutableArray<byte> bytes = module.Module.PEReaderOpt.GetEntireImage().GetContent();
-                    ModuleData moduleData;
-                    if (isManifestModule)
-                    {
-                        fullNameSet.Add(identity.GetDisplayName());
-                        moduleData = new ModuleData(identity,
-                                                    OutputKind.DynamicallyLinkedLibrary,
-                                                    bytes,
-                                                    pdb: default(ImmutableArray<byte>),
-                                                    inMemoryModule: true);
-                    }
-                    else
-                    {
-                        moduleData = new ModuleData(module.Name,
-                                                    bytes,
-                                                    pdb: default(ImmutableArray<byte>),
-                                                    inMemoryModule: true);
-                    }
-
-                    dependencies.Add(moduleData);
-                    isManifestModule = false;
-                }
-            }
-        }
-
-        private static IEnumerable<ModuleMetadata> EnumerateModules(Metadata metadata)
-        {
-            return (metadata.Kind == MetadataImageKind.Assembly) ? ((AssemblyMetadata)metadata).GetModules().AsEnumerable() : SpecializedCollections.SingletonEnumerable((ModuleMetadata)metadata);
-        }
-
-        private static EmitOutput? EmitCompilation(
-            Compilation compilation,
-            IEnumerable<ResourceDescription> manifestResources,
-            List<ModuleData> dependencies,
-            DiagnosticBag diagnostics,
-            CompilationTestData testData
-        )
-        {
-            // A Compilation can appear multiple times in a depnedency graph as both a Compilation and as a MetadataReference
-            // value.  Iterate the Compilations eagerly so they are always emitted directly and later references can re-use 
-            // the value.  This gives better, and consistent, diagostic information.
-            var referencedCompilations = FindReferencedCompilations(compilation);
-            var fullNameSet = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-
-            foreach (var referencedCompilation in referencedCompilations)
-            {
-                var emitData = EmitCompilationCore(referencedCompilation, null, diagnostics, null);
-                if (emitData.HasValue)
-                {
-                    var moduleData = new ModuleData(referencedCompilation.Assembly.Identity,
-                                                    OutputKind.DynamicallyLinkedLibrary,
-                                                    emitData.Value.Assembly,
-                                                    pdb: default(ImmutableArray<byte>),
-                                                    inMemoryModule: true);
-                    fullNameSet.Add(moduleData.Id.FullName);
-                    dependencies.Add(moduleData);
-                }
-            }
-
-            // Now that the Compilation values have been emitted, emit the non-compilation references
-            foreach (var current in (new[] { compilation }).Concat(referencedCompilations))
-            {
-                EmitReferences(current, fullNameSet, dependencies, diagnostics);
-            }
-
-            return EmitCompilationCore(compilation, manifestResources, diagnostics, testData);
-        }
-
-        private static EmitOutput? EmitCompilationCore(
-            Compilation compilation,
-            IEnumerable<ResourceDescription> manifestResources,
-            DiagnosticBag diagnostics,
-            CompilationTestData testData
-        )
-        {
-            using (var executableStream = new MemoryStream())
-            {
-                var pdb = default(ImmutableArray<byte>);
-                var assembly = default(ImmutableArray<byte>);
-                var pdbStream = MonoHelpers.IsRunningOnMono()
-                    ? null
-                    : new MemoryStream();
-
-                EmitResult result;
-                try
-                {
-                    result = compilation.Emit(
-                        executableStream,
-                        pdbStream: pdbStream,
-                        xmlDocumentationStream: null,
-                        win32Resources: null,
-                        manifestResources: manifestResources,
-                        options: EmitOptions.Default,
-                        debugEntryPoint: null,
-                        testData: testData,
-                        getHostDiagnostics: null,
-                        cancellationToken: default(CancellationToken));
-                }
-                finally
-                {
-                    if (pdbStream != null)
-                    {
-                        pdb = pdbStream.ToImmutable();
-                        pdbStream.Dispose();
-                    }
-                }
-
-                diagnostics.AddRange(result.Diagnostics);
-                assembly = executableStream.ToImmutable();
-
-                if (result.Success)
-                {
-                    return new EmitOutput(assembly, pdb);
-                }
-
-                return null;
-            }
-        }
-
         public void Emit(
             Compilation mainCompilation,
             IEnumerable<ResourceDescription> manifestResources,
@@ -403,7 +194,7 @@ namespace Microsoft.CodeAnalysis.Test.Utilities
 
             var diagnostics = DiagnosticBag.GetInstance();
             var dependencies = new List<ModuleData>();
-            var mainOutput = EmitCompilation(mainCompilation, manifestResources, dependencies, diagnostics, _testData);
+            var mainOutput = RuntimeUtilities.EmitCompilation(mainCompilation, manifestResources, dependencies, diagnostics, _testData);
 
             _emitData = new EmitData();
             _emitData.Diagnostics = diagnostics.ToReadOnlyAndFree();
@@ -484,7 +275,7 @@ namespace Microsoft.CodeAnalysis.Test.Utilities
             return _emitData;
         }
 
-        internal ImmutableArray<Diagnostic> GetDiagnostics()
+        public ImmutableArray<Diagnostic> GetDiagnostics()
         {
             return GetEmitData().Diagnostics;
         }
@@ -499,7 +290,7 @@ namespace Microsoft.CodeAnalysis.Test.Utilities
             return GetEmitData().MainModulePdb;
         }
 
-        internal IList<ModuleData> GetAllModuleData()
+        public IList<ModuleData> GetAllModuleData()
         {
             return GetEmitData().AllModuleData;
         }
@@ -511,14 +302,14 @@ namespace Microsoft.CodeAnalysis.Test.Utilities
             emitData.Manager.PeVerifyModules(new[] { emitData.MainModule.FullName });
         }
 
-        internal string[] PeVerifyModules(string[] modulesToVerify, bool throwOnError = true)
+        public string[] PeVerifyModules(string[] modulesToVerify, bool throwOnError = true)
         {
             var emitData = GetEmitData();
             emitData.RuntimeData.PeverifyRequested = true;
             return emitData.Manager.PeVerifyModules(modulesToVerify, throwOnError);
         }
 
-        internal SortedSet<string> GetMemberSignaturesFromMetadata(string fullyQualifiedTypeName, string memberName)
+        public SortedSet<string> GetMemberSignaturesFromMetadata(string fullyQualifiedTypeName, string memberName)
         {
             var emitData = GetEmitData();
             var searchIds = emitData.AllModuleData.Select(x => x.Id).ToList();
@@ -554,7 +345,7 @@ namespace Microsoft.CodeAnalysis.Test.Utilities
             _disposed = true;
         }
 
-        internal CompilationTestData GetCompilationTestData()
+        CompilationTestData IInternalRuntimeUtility.GetCompilationTestData()
         {
             if (_testData.Module == null)
             {
