+++ src/Test/Utilities/Desktop/MetadataSignatureUnitTestHelper.cs
@@ -15,15 +15,15 @@ namespace Microsoft.CodeAnalysis.Test.Utilities
         public string ExpectedSignature { get; set; }
     }
 
-    public class MetadataSignatureUnitTestHelper
+    internal class MetadataSignatureUnitTestHelper
     {
         /// <summary>
         /// Uses Reflection to verify that the specified member signatures are present in emitted metadata
         /// </summary>
         /// <param name="appDomainHost">Unit test AppDomain host</param>
         /// <param name="expectedSignatures">Baseline signatures - use the Signature() factory method to create instances of SignatureDescription</param>
-        public static void VerifyMemberSignatures(
-            HostedRuntimeEnvironment appDomainHost, params SignatureDescription[] expectedSignatures)
+        internal static void VerifyMemberSignatures(
+            IRuntimeUtility appDomainHost, params SignatureDescription[] expectedSignatures)
         {
             Assert.NotNull(expectedSignatures);
             Assert.NotEmpty(expectedSignatures);
@@ -76,7 +76,7 @@ namespace Microsoft.CodeAnalysis.Test.Utilities
         /// <param name="actualSignatures">List of found signatures matching member name</param>
         /// <returns>True if a matching member signature was found, false otherwise</returns>
         private static bool VerifyMemberSignatureHelper(
-            HostedRuntimeEnvironment appDomainHost, string fullyQualifiedTypeName, string memberName,
+            IRuntimeUtility appDomainHost, string fullyQualifiedTypeName, string memberName,
             ref string expectedSignature, out List<string> actualSignatures)
         {
             Assert.False(string.IsNullOrWhiteSpace(fullyQualifiedTypeName), "'fullyQualifiedTypeName' can't be null or empty");
