+++ src/Test/Utilities/Portable/TestUtilities.csproj
@@ -115,6 +115,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="CommonTestBase.cs" />
+    <Compile Include="CoreCLRRuntimeEnvironment.cs" />
+    <Compile Include="Exceptions.cs" />
     <Compile Include="ICompilationVerifier.cs" />
     <Compile Include="MetadataSignatureUnitTestHelper.cs" />
     <Compile Include="ModuleData.cs" />
