+++ src/Test/Utilities/Shared/Compilation/IRuntimeUtility.cs
@@ -1,0 +1,253 @@
+﻿
+using System;
+using System.Collections.Generic;
+using System.Collections.Immutable;
+using System.IO;
+using System.Linq;
+using System.Threading;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CodeGen;
+using Microsoft.CodeAnalysis.Emit;
+using Microsoft.CodeAnalysis.Test.Utilities;
+using Roslyn.Utilities;
+
+namespace Roslyn.Test.Utilities
+{
+    public class  RuntimeUtilityFactory
+    {
+        internal static IRuntimeUtility NewInstance(IEnumerable<ModuleData> additionalDependencies = null)
+    {
+
+#if DNX
+            return new CoreCLRRuntimeEnvironment(additionalDependencies);
+#else
+            return new HostedRuntimeEnvironment(additionalDependencies);
+#endif
+        }
+    }
+
+    internal struct EmitOutput
+    {
+        internal ImmutableArray<byte> Assembly { get; }
+        internal ImmutableArray<byte> Pdb { get; }
+
+        internal EmitOutput(ImmutableArray<byte> assembly, ImmutableArray<byte> pdb)
+        {
+            Assembly = assembly;
+            Pdb = pdb;
+        }
+    }
+
+    internal static class RuntimeUtilities
+    {
+        private static IEnumerable<ModuleMetadata> EnumerateModules(Metadata metadata)
+        {
+            return (metadata.Kind == MetadataImageKind.Assembly) ? ((AssemblyMetadata)metadata).GetModules().AsEnumerable() : SpecializedCollections.SingletonEnumerable((ModuleMetadata)metadata);
+        }
+
+        /// <summary>
+        /// Emit all of the references which are not directly or indirectly a <see cref="Compilation"/> value.
+        /// </summary>
+        internal static void EmitReferences(Compilation compilation, HashSet<string> fullNameSet, List<ModuleData> dependencies, DiagnosticBag diagnostics)
+        {
+            // NOTE: specifically don't need to consider previous submissions since they will always be compilations.
+            foreach (var metadataReference in compilation.References)
+            {
+                if (metadataReference is CompilationReference)
+                {
+                    continue;
+                }
+
+                var peRef = (PortableExecutableReference)metadataReference;
+                var metadata = peRef.GetMetadata();
+                var isManifestModule = peRef.Properties.Kind == MetadataImageKind.Assembly;
+                var identity = isManifestModule
+                    ? ((AssemblyMetadata)metadata).GetAssembly().Identity
+                    : null;
+
+                // If this is an indirect reference to a Compilation then it is already been emitted 
+                // so no more work to be done.
+                if (isManifestModule && fullNameSet.Contains(identity.GetDisplayName()))
+                {
+                    continue;
+                }
+
+                foreach (var module in EnumerateModules(metadata))
+                {
+                    ImmutableArray<byte> bytes = module.Module.PEReaderOpt.GetEntireImage().GetContent();
+                    ModuleData moduleData;
+                    if (isManifestModule)
+                    {
+                        fullNameSet.Add(identity.GetDisplayName());
+                        moduleData = new ModuleData(identity,
+                                                    OutputKind.DynamicallyLinkedLibrary,
+                                                    bytes,
+                                                    pdb: default(ImmutableArray<byte>),
+                                                    inMemoryModule: true);
+                    }
+                    else
+                    {
+                        moduleData = new ModuleData(module.Name,
+                                                    bytes,
+                                                    pdb: default(ImmutableArray<byte>),
+                                                    inMemoryModule: true);
+                    }
+
+                    dependencies.Add(moduleData);
+                    isManifestModule = false;
+                }
+            }
+        }
+        /// <summary>
+        /// Find all of the <see cref="Compilation"/> values reachable from this instance.
+        /// </summary>
+        /// <param name="compilation"></param>
+        /// <returns></returns>
+        private static List<Compilation> FindReferencedCompilations(Compilation original)
+        {
+            var list = new List<Compilation>();
+            var toVisit = new Queue<Compilation>(FindDirectReferencedCompilations(original));
+
+            while (toVisit.Count > 0)
+            {
+                var current = toVisit.Dequeue();
+                if (list.Contains(current))
+                {
+                    continue;
+                }
+
+                list.Add(current);
+
+                foreach (var other in FindDirectReferencedCompilations(current))
+                {
+                    toVisit.Enqueue(other);
+                }
+            }
+
+            return list;
+        }
+
+        private static List<Compilation> FindDirectReferencedCompilations(Compilation compilation)
+        {
+            var list = new List<Compilation>();
+            var previousCompilation = compilation.ScriptCompilationInfo?.PreviousScriptCompilation;
+            if (previousCompilation != null)
+            {
+                list.Add(previousCompilation);
+            }
+
+            foreach (var reference in compilation.References.OfType<CompilationReference>())
+            {
+                list.Add(reference.Compilation);
+            }
+
+            return list;
+        }
+        internal static EmitOutput? EmitCompilation(
+            Compilation compilation,
+            IEnumerable<ResourceDescription> manifestResources,
+            List<ModuleData> dependencies,
+            DiagnosticBag diagnostics,
+            CompilationTestData testData
+        )
+        {
+            // A Compilation can appear multiple times in a depnedency graph as both a Compilation and as a MetadataReference
+            // value.  Iterate the Compilations eagerly so they are always emitted directly and later references can re-use 
+            // the value.  This gives better, and consistent, diagostic information.
+            var referencedCompilations = FindReferencedCompilations(compilation);
+            var fullNameSet = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
+
+            foreach (var referencedCompilation in referencedCompilations)
+            {
+                var emitData = RuntimeUtilities.EmitCompilationCore(referencedCompilation, null, diagnostics, null);
+                if (emitData.HasValue)
+                {
+                    var moduleData = new ModuleData(referencedCompilation.Assembly.Identity,
+                                                    OutputKind.DynamicallyLinkedLibrary,
+                                                    emitData.Value.Assembly,
+                                                    pdb: default(ImmutableArray<byte>),
+                                                    inMemoryModule: true);
+                    fullNameSet.Add(moduleData.Id.FullName);
+                    dependencies.Add(moduleData);
+                }
+            }
+
+            // Now that the Compilation values have been emitted, emit the non-compilation references
+            foreach (var current in (new[] { compilation }).Concat(referencedCompilations))
+            {
+                EmitReferences(current, fullNameSet, dependencies, diagnostics);
+            }
+
+            return RuntimeUtilities.EmitCompilationCore(compilation, manifestResources, diagnostics, testData);
+        }
+
+        internal static EmitOutput? EmitCompilationCore(
+            Compilation compilation,
+            IEnumerable<ResourceDescription> manifestResources,
+            DiagnosticBag diagnostics,
+            CompilationTestData testData
+        )
+        {
+            using (var executableStream = new MemoryStream())
+            {
+                var pdb = default(ImmutableArray<byte>);
+                var assembly = default(ImmutableArray<byte>);
+                var pdbStream = MonoHelpers.IsRunningOnMono()
+                    ? null
+                    : new MemoryStream();
+
+                EmitResult result;
+                try
+                {
+                    result = compilation.Emit(
+                        executableStream,
+                        pdbStream: pdbStream,
+                        xmlDocumentationStream: null,
+                        win32Resources: null,
+                        manifestResources: manifestResources,
+                        options: EmitOptions.Default,
+                        debugEntryPoint: null,
+                        testData: testData,
+                        getHostDiagnostics: null,
+                        cancellationToken: default(CancellationToken));
+                }
+                finally
+                {
+                    if (pdbStream != null)
+                    {
+                        pdb = pdbStream.ToImmutable();
+                        pdbStream.Dispose();
+                    }
+                }
+
+                diagnostics.AddRange(result.Diagnostics);
+                assembly = executableStream.ToImmutable();
+
+                if (result.Success)
+                {
+                    return new EmitOutput(assembly, pdb);
+                }
+
+                return null;
+            }
+        }
+    }
+
+    public interface IRuntimeUtility: System.IDisposable
+    {
+        void Emit(Microsoft.CodeAnalysis.Compilation mainCompilation, IEnumerable<ResourceDescription> manifestResources, bool usePdbForDebugging = false);
+        int Execute(string moduleName, string expectedOutput);
+        ImmutableArray<byte> GetMainImage();
+        ImmutableArray<byte> GetMainPdb();
+        ImmutableArray<Diagnostic> GetDiagnostics();
+        SortedSet<string> GetMemberSignaturesFromMetadata(string fullyQualifiedTypeName, string memberName);
+        IList<ModuleData> GetAllModuleData();
+        void PeVerify();
+        string[] PeVerifyModules(string[] modulesToVerify, bool throwOnError = true);
+    }
+
+    internal interface IInternalRuntimeUtility
+    {
+        CompilationTestData GetCompilationTestData();
+    }
+}
